Fixed some kata I had messed up that the tests found and added one for today
